src/signupservice/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==37.0.4
Deprecated==1.2.13
firebase-admin==6.0.1
Flask==2.2.2
flatbuffers==1.12
flightsql-dbapi==0.2.0
gast==0.4.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.1
greenlet==2.0.2
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.13.0
h5py==3.6.0
httplib2==0.21.0
idna==3.3
influxdb==5.3.1
influxdb-client==1.36.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
keras-nightly==2.10.0.dev2022042007
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.2
msgpack==1.0.4
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
proto-plus==1.22.1
protobuf==4.21.12
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pygal==3.0.0
pygal-maps-ch==1.0.1
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrebase5==5.0.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.7.1
reactivex==4.0.4
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.8
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
tb-nightly==2.9.0a20220420
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
# tf-estimator-nightly==2.10.0.dev2022042008
# tf-nightly==2.10.0.dev20220420
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
# vgamepad==0.0.8
Werkzeug==2.2.3
wrapt==1.14.0
wsproto==1.1.0
